refactor(replsession): extract persistence helpers from service
